highlighter updates
